backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.6.3
python-dotenv==1.0.1
pydantic==2.6.4
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1